depthai-niryo/deploy/requirements.txt
depthai
argparse
uvicorn
json
pathlib
fastapi
paho-mqtt
asyncio
jsonpickle 
catkin_pkg 
setuptools
Cython 
pyparsing 
numpy 
